# docs/swagger.yaml
basePath: /
definitions:
  controllers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  controllers.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
host: auth-service-byz1bp-c95241-204-12-199-113.traefik.me
info:
  contact: {}
  description: Documentação da API de autenticação
  title: Auth Service API
  version: "1.0"
paths:
  /api/profile:
    get:
      description: Requer um token JWT válido no header Authorization
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Retorna o perfil do usuário autenticado
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentica usuário e retorna um token JWT
      parameters:
      - description: Credenciais do usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Realiza login de um usuário
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Registra um usuário com nome, email e senha
      parameters:
      - description: Dados do usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cadastra um novo usuário
      tags:
      - Auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"